requirements.txt
asgiref==3.4.1
Babel==2.9.1
Django==3.2.11
django-money==2.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
py-moneyed==1.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
